✨ feat: 마이페이지 이름변경, 페이지 서버로 변경
